refactor(usage): share usage history normalizer
